chore: decompose modules and reuse import logic for expanding files
License: MIT
